Rename request to MappedParameter
